[Refactor] #132 - 작품 평가 화면 파일 분리 및 기간 시트 ViewModel 도입
